+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
  - General-purpose BgzipWriter
+ - BgzipFastaWriter and BgzipFastqWriter
 
 ## [1.0.3] - 2019-05-20